Kernel: Move ProcessTracer into a new Tracing/ directory
We're going to have more tracing-related files soon, so let's group them.
